Implement file chooser and resolution strategy (#1)
Sort out the confusion between index and archive in the model. Formalize
current resolution strategy by preparing to introduce a resolution
pattern. A resolution strategy essentially takes an input and returns a
result pointing to an archive and index file. Validation can be added to
a `resolve` method. Currently we only support loading an archive next to
an index file.

A file chooser helps pick files from a the local machine, instead of
hardcoded paths.