Rudimentary markup support (#3)
* Parse wiki page markup for exact match from archive

If we can find an exact match of the user's query in our index we parse
the page from the archive.

Known issues:
- Parsing the archive is not optimized (e.g. we probably read more than
  we have to).

* Markup: rudimentary support

Convert the wiki markup into tokens that can in turn be converted into
JavaFX components. We are using the TextFow layout to combine the
components into a single text flow.

* Markup: Add support for links

Parse links and attempt to also cover the "word ending" feature, where a
link points to a term, but additional characters need to be considered.
E.g. "[[Word]]ending" should result in the text "Wordending" and link
to the term "Word".