Markup: Add support for indents
Wikitonary text heavily use the indent feature, e.g.:

`: Some text that should be indented.`

For a better user experience we are adding support for indent.

The MarkupParser implementation is cleaned up a bit and additional
helper functionality is added.

This change also switches parsing for paragraph/text to be newline
delimited. That just simplifies keeping indentation markup and the use
of colons in text separate.

   Because we attempt to parse indentation markup before regular text,
   we can simply test if the next character is a colon. If text is
   terminated by a newline we can assume to always find indentation at
   the beginning of the line. Text parsing doesn't need to worry about
   figuring out if a ':' terminates the current text token.