Markup: Add support for italics
Introduce a `TextContent` parser, as outlined in the grammar. Ensure
that the text parser doesn't return a TextToken with an empty value.
Otherwise we'll hang in an endless loop as the text parser can't advance
beyond the final italics delimiter.

In a first attempt we pollute the WikiTextFlow class with a flag to turn
on italic text. In the long run this should be migrated to it's own
class.