Parser: improve parsing of word-ending links
Previously parsing word-ending links would expect a space to "terminate"
the label part. However sometimes links with word-endings are wrapped in
italics:

    ''example [[Link]]s'' more text

In that case the italics parser would correctly encounter the link, but
the link parser would parse `Links''` as label. When control is returned
to the italics parser, it can't find the terminating characters for the
italics anymore and we'd never create the correct token.