Markup parser improvements (#10)
* Parser: improve parsing of word-ending links

Previously parsing word-ending links would expect a space to "terminate"
the label part. However sometimes links with word-endings are wrapped in
italics:

    ''example [[Link]]s'' more text

In that case the italics parser would correctly encounter the link, but
the link parser would parse `Links''` as label. When control is returned
to the italics parser, it can't find the terminating characters for the
italics anymore and we'd never create the correct token.

* Implement a fuzzer to check for parsing errors

It's tedious to manually search for Wiktionary entries that break the
parser. The new FuzzerApp takes an index and tries to parse most entries
from the corresponding archive. Eventually a new index file is written
containing the entries that cause exceptions. This helps keeping the
feedback loop tight.

* Parser: Skip media embeddings

Adds a SkipToken to signal content that should be skipped or produces no
output. Media embeddings look a lot like links which caused the link
parser to never produced correct output. Which in turn prevents some
terms from rendering.

Smaller fixes and improvements around parsing links.